pyproject.toml
[project]
name = "ic-urfu-odoo-module"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pydantic-settings>=2.13.1",
    "python-docx>=1.2.0",
    "structlog>=25.5.0",
]

[dependency-groups]
dev = [
    "ruff>=0.15.8",
]
